Display Two Seven Segment Displays by Transistor Multiplexer
The main file is Digital_IO.c where code-friendly port initialization, system timer, and state multiplexing occurs.

Interfaced NPN transistor, 7-Segment display, and Input Switch with consideration to their datasheets and electrical properties.

De-bounced the switch, so on a button press, COUNT increments and that number 0-99 are sent to the displays.

To decrease the total pins required for the integration, 7 display-data pins are in parallel with the hex displays. Two transistors "base" control pins are used for toggling the two hex displays at a 10ms frequency.

The outcome is a perceived solid display even though the displays are alternating rapidly.